refactor: separate 3x and 4x basics
